+++ src/date.rs
@@ -9,6 +9,7 @@ use core::cmp::Ordering;
 use core::ops::{Add, Sub};
 use core::{fmt, hash};
 use oldtime::Duration as OldDuration;
+use std::convert::TryFrom;
 
 #[cfg(feature = "unstable-locales")]
 use format::Locale;
@@ -271,6 +272,19 @@ impl<Tz: TimeZone> Date<Tz> {
     pub fn naive_local(&self) -> NaiveDate {
         self.date
     }
+
+    /// Retrieve the elapsed years from now to the given [`Date`].
+    pub fn elapsed_years(&self) -> u32 {
+        let now = Utc::today().with_timezone(&self.timezone());
+
+        let years = if (now.month(), now.day()) < (self.month(), self.day()) {
+            now.year() - self.year() - 1
+        } else {
+            now.year() - self.year()
+        };
+
+        u32::try_from(years).unwrap_or(0)
+    }
 }
 
 /// Maps the local date to other date with given conversion function.
@@ -492,5 +506,25 @@ where
 {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{}{}", self.naive_local(), self.offset)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use consts::f64;
+    use offset::Utc;
+    use oldtime::Duration;
+
+    #[test]
+    fn test_years_elapsed() {
+        // This is always at least one year because 1 year = 52.1775 weeks.
+        let one_year_ago = Utc::today() - Duration::weeks((f64::WEEK_PER_YEAR * 1.5).ceil() as i64);
+        // A bit more than 2 years.
+        let two_year_ago = Utc::today() - Duration::weeks((f64::WEEK_PER_YEAR * 2.5).ceil() as i64);
+
+        assert_eq!(one_year_ago.elapsed_years(), 1);
+        assert_eq!(two_year_ago.elapsed_years(), 2);
+        // if the given Date is later than now, the function will always return 0.
+        assert_eq!((Utc::today() + Duration::weeks(12)).elapsed_years(), 0);
     }
 }